Changes in the writer filename generated. Also added a external directory to share files between the producer and the consumer. Changed the size of the bitmap, according to the new specifications.

+++ producer/src/Interfaces/IFileWriter.cs
@@ -1,8 +1,8 @@
 ﻿namespace Producer
 {
-    interface IFileWriter
-    {
-        void InitializeDirectory(string directory);
-        void Write(string file, double[] data);
-    }
+	interface IFileWriter
+	{
+		void InitializeDirectory(string directory);
+		void Write(string file, float[] data);
+	}
 }
